# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BookReviver",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "description": "List the projects of the signed-in account, most recently updated first.",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Create a project from the description of its book, and point to it in the ``Location`` header.",
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/stages/{stage}/recipe": {
      "get": {
        "tags": [
          "processing"
        ],
        "summary": "Get Recipe",
        "description": "Return the active recipe of a stage, which a project creates the first time the stage is asked for.",
        "operationId": "get_recipe_api_v1_projects__project_id__stages__stage__recipe_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the project",
              "title": "Project Id"
            },
            "description": "Identifier of the project"
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Stage",
              "description": "Stage of the pipeline"
            },
            "description": "Stage of the pipeline"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "processing"
        ],
        "summary": "Put Recipe",
        "description": "Replace the name and the steps of the active recipe, which marks the pages it processed stale.\n\nNo page is processed again by this request; the stage is run by ``POST .../run``. A step whose processor is unknown\nor of another stage, or whose parameters do not fit, answers 422.",
        "operationId": "put_recipe_api_v1_projects__project_id__stages__stage__recipe_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the project",
              "title": "Project Id"
            },
            "description": "Identifier of the project"
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Stage",
              "description": "Stage of the pipeline"
            },
            "description": "Stage of the pipeline"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/stages/{stage}/variants/{recipe_id}": {
      "put": {
        "tags": [
          "processing"
        ],
        "summary": "Put Variant",
        "description": "Replace the name and the steps of a recipe of a stage, which marks the pages it processed stale.",
        "operationId": "put_variant_api_v1_projects__project_id__stages__stage__variants__recipe_id__put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the project",
              "title": "Project Id"
            },
            "description": "Identifier of the project"
          },
          {
            "name": "stage",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Stage",
              "description": "Stage of the pipeline"
            },
            "description": "Stage of the pipeline"
          },
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the recipe",
              "title": "Recipe Id"
            },
            "description": "Identifier of the recipe"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Problem title"
          },
          "type": {
            "type": "string",
            "title": "Problem type"
          },
          "status": {
            "type": "integer",
            "title": "Status code"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "errors"
        ],
        "title": "RequestValidationError"
      },
...
      "Page_ProjectSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[ProjectSchema]"
      },
...
      "Stage": {
        "type": "string",
        "enum": [
          "import",
          "page-split",
          "page-order",
          "geometry",
          "cleanup",
          "layout",
          "background",
          "recognition",
          "proofreading",
          "typesetting"
        ],
        "title": "Stage",
        "description": "A step of the digitisation pipeline, in pipeline order."
      },
...
      "Problem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Problem title"
          },
          "type": {
            "type": "string",
            "title": "Problem type"
          },
          "status": {
            "type": "integer",
            "title": "Status code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem detail"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bookreviver_session"
      }
    }
  }
}